+++ skeleton/modules/engine/main_engine.h
@@ -7,6 +7,6 @@
 
 void main_engine_init();
 
-void main_engine_vals(i16 left, i16 right);
+void main_engine_vals(i32 left, i32 right);
 
 #endif //MAIN_ENGINE_H
