+++ skeleton/modules/engine/turret_rotation.h
@@ -6,6 +6,6 @@
 #include "shared_config.h"
 
 void turret_rotation_init();
-void turret_rotation_rotate(i16 val);
+void turret_rotation_rotate(i32 val);
 
 #endif //TURRET_ROTATION_H
